agent/requirements.txt
Flask==2.3.3
PyYAML==6.0.1
psutil==5.9.5
requests==2.31.0
numpy==1.24.3
gunicorn==21.2.0
Werkzeug==2.3.7
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2

# Development dependencies
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
black==23.7.0
flake8==6.0.0
mypy==1.5.1